feat: adding trt8 supporting models. And fix path not found error.
